@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Razor">
 
 	<PropertyGroup>
-		<TargetFramework>net9.0</TargetFramework>
+		<TargetFramework>net10.0</TargetFramework>
 		<Nullable>enable</Nullable>
 		<ImplicitUsings>enable</ImplicitUsings>
 	</PropertyGroup>
@@ -11,7 +11,7 @@
 	</ItemGroup>
 
 	<ItemGroup>
-		<PackageReference Include="Microsoft.AspNetCore.Components.Web" Version="9.0.9" />
+		<PackageReference Include="Microsoft.AspNetCore.Components.Web" Version="10.0.1" />
 	</ItemGroup>
 
 	<ItemGroup>
@@ -28,8 +28,7 @@
 	</PropertyGroup>
 
 	<ItemGroup Condition="'$(Configuration)' == 'Debug'">
-	  <ProjectReference Include="..\BlazorSchool.BlazorLibrary2.Generators\BlazorSchool.BlazorLibrary2.Generators.csproj" OutputItemType="Analyzer"
-                    ReferenceOutputAssembly="false"/>
+	  <ProjectReference Include="..\BlazorSchool.BlazorLibrary2.Generators\BlazorSchool.BlazorLibrary2.Generators.csproj" OutputItemType="Analyzer" ReferenceOutputAssembly="false" />
 	</ItemGroup>
 
 </Project>
